+++ packages/gitbook/src/components/Embeddable/EmbeddableRootLayout.tsx
@@ -12,6 +12,7 @@ import { SpaceLayoutServerContext } from '../SpaceLayout';
 import { Trademark } from '../TableOfContents/Trademark';
 import { NavigationLoader } from '../primitives/NavigationLoader';
 import { EmbeddableIframeAPI } from './EmbeddableIframeAPI';
+import { IfEmbeddableTrademark } from './EmbeddableTrademark';
 
 type EmbeddableRootLayoutProps = {
     context: GitBookSiteContext;
@@ -58,11 +59,13 @@ export async function EmbeddableRootLayout({
                         <div className="fixed inset-0 flex flex-col">
                             {children}
                             {context.customization.trademark.enabled ? (
-                                <Trademark
-                                    className="rounded-none! border-x-0 border-t border-b-0 bg-tint-solid/1 depth-flat:bg-tint-solid/1 px-4 py-2.5 text-tint/8"
-                                    context={context}
-                                    placement={SiteInsightsTrademarkPlacement.Embed}
-                                />
+                                <IfEmbeddableTrademark>
+                                    <Trademark
+                                        className="rounded-none! border-x-0 border-t border-b-0 bg-tint-solid/1 depth-flat:bg-tint-solid/1 px-4 py-2.5 text-tint/8"
+                                        context={context}
+                                        placement={SiteInsightsTrademarkPlacement.Embed}
+                                    />
+                                </IfEmbeddableTrademark>
                             ) : null}
                         </div>
                         <EmbeddableIframeAPI